+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/api/MapboxRouteLineApi.kt
@@ -29,6 +29,7 @@ import com.mapbox.navigation.core.trip.session.RouteProgressObserver
 import com.mapbox.navigation.ui.base.util.MapboxNavigationConsumer
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils
 import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils.getMatchingColors
+import com.mapbox.navigation.ui.maps.internal.route.line.RouteLineTrimOffset
 import com.mapbox.navigation.ui.maps.route.RouteLayerConstants
 import com.mapbox.navigation.ui.maps.route.line.model.ClosestRouteValue
 import com.mapbox.navigation.ui.maps.route.line.model.ExtractedRouteData
@@ -1421,7 +1422,10 @@ class MapboxRouteLineApi(
                         primaryRouteBaseExpressionProducer,
                         primaryRouteCasingExpressionProducer,
                         primaryRouteTrafficLineExpressionProducer,
-                        primaryRouteRestrictedSectionsExpressionProducer
+                        primaryRouteRestrictedSectionsExpressionProducer,
+                        RouteLineTrimOffset(
+                            routeLineOptions.vanishingRouteLine?.vanishPointOffset ?: 0.0
+                        )
                     )
                 ),
                 alternativeRouteLinesData = listOf(