+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/model/RouteLineDynamicData.kt
@@ -1,5 +1,7 @@
 package com.mapbox.navigation.ui.maps.route.line.model
 
+import com.mapbox.navigation.ui.maps.internal.route.line.RouteLineTrimOffset
+
 /**
  * Provides information needed to draw a route.
  *
@@ -12,7 +14,8 @@ class RouteLineDynamicData internal constructor(
     val baseExpressionProvider: RouteLineExpressionProvider,
     val casingExpressionProvider: RouteLineExpressionProvider,
     val trafficExpressionProvider: RouteLineExpressionProvider?,
-    val restrictedSectionExpressionProvider: RouteLineExpressionProvider?
+    val restrictedSectionExpressionProvider: RouteLineExpressionProvider?,
+    internal val trimOffset: RouteLineTrimOffset? = null
 ) {
 
     /**
@@ -22,7 +25,8 @@ class RouteLineDynamicData internal constructor(
         baseExpressionProvider,
         casingExpressionProvider,
         trafficExpressionProvider,
-        restrictedSectionExpressionProvider
+        restrictedSectionExpressionProvider,
+        trimOffset
     )
 
     /**
@@ -32,12 +36,14 @@ class RouteLineDynamicData internal constructor(
      * @param casingExpressionProvider expression used to style the case of the line
      * @param trafficExpressionProvider expression used to style the congestion colors on the line
      * @param restrictedSectionExpressionProvider expression used to style the restricted sections on the line
+     * @param trimOffset a value representing the section of the line that should be trimmed and made transparent
      */
     class MutableRouteLineDynamicData internal constructor(
         var baseExpressionProvider: RouteLineExpressionProvider,
         var casingExpressionProvider: RouteLineExpressionProvider,
         var trafficExpressionProvider: RouteLineExpressionProvider?,
-        var restrictedSectionExpressionProvider: RouteLineExpressionProvider?
+        var restrictedSectionExpressionProvider: RouteLineExpressionProvider?,
+        internal var trimOffset: RouteLineTrimOffset? = null
     ) {
 
         /**
@@ -47,7 +53,8 @@ class RouteLineDynamicData internal constructor(
             baseExpressionProvider,
             casingExpressionProvider,
             trafficExpressionProvider,
-            restrictedSectionExpressionProvider
+            restrictedSectionExpressionProvider,
+            trimOffset
         )
     }
 
@@ -65,6 +72,7 @@ class RouteLineDynamicData internal constructor(
         if (trafficExpressionProvider != other.trafficExpressionProvider) return false
         if (restrictedSectionExpressionProvider != other.restrictedSectionExpressionProvider)
             return false
+        if (trimOffset != other.trimOffset) return false
 
         return true
     }
@@ -77,6 +85,7 @@ class RouteLineDynamicData internal constructor(
         result = 31 * result + casingExpressionProvider.hashCode()
         result = 31 * result + (trafficExpressionProvider?.hashCode() ?: 0)
         result = 31 * result + (restrictedSectionExpressionProvider?.hashCode() ?: 0)
+        result = 31 * result + (trimOffset?.hashCode() ?: 0)
         return result
     }
 
@@ -88,7 +97,8 @@ class RouteLineDynamicData internal constructor(
             "baseExpressionProvider=$baseExpressionProvider, " +
             "casingExpressionProvider=$casingExpressionProvider, " +
             "trafficExpressionProvider=$trafficExpressionProvider, " +
-            "restrictedSectionExpressionProvider=$restrictedSectionExpressionProvider" +
+            "restrictedSectionExpressionProvider=$restrictedSectionExpressionProvider," +
+            "trimOffset=$trimOffset" +
             ")"
     }
 }